Restructure menu components
